Update the environment libraries to recent versions. This required changing the way rngs are passed to the model. It also exposed a bug in MCTS, that has been fixed.

+++ omega/minihack/utils.py
@@ -6,7 +6,7 @@ def glyph_pos(glyphs, glyph):
     assert len(glyph_positions) == 2
     if glyph_positions[0].shape[0] == 0:
         return None
-    return np.array([glyph_positions[0][0], glyph_positions[1][0]], dtype=np.float)
+    return np.array([glyph_positions[0][0], glyph_positions[1][0]], dtype=np.float32)
 
 
 def print_char_glyphs(char_glyphs):
